refactor(editor): table paste 모델을 분리한다
